% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{calculate_ref_mean}
\alias{calculate_ref_mean}
\title{Calculate reference mean of  samples}
\usage{
calculate_ref_mean(df_qc)
}
\arguments{
\item{df_qc}{peak matrix of QC samples}
}
\value{
vector of reference mean values
}
\description{
Calculate reference mean of  samples
}
