% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPeakMatrixOrientation.R
\name{check_peak_matrix_orientation}
\alias{check_peak_matrix_orientation}
\title{Check if peak matrix is in format features in rows, samples in columns.}
\usage{
check_peak_matrix_orientation(peak_data, classes = NULL)
}
\arguments{
\item{peak_data}{peak matrix}

\item{classes}{Vector of class labels}
}
\value{
matrix where samples are represented in columns and features in rows
}
\description{
All functions in pmp pacakge expect input peak matrix to have samples as columns and
measured features in rows. This function will check input matrix orientation and will
transpose it if needed. If class labels are provided this function will check if the 
length of labels matches number of samples in peak matrix.
}
\examples{

attach (testData)
out <- check_peak_matrix_orientation (peak_data=t(testData$data))

}
