% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPeakMatrixOrientation.R
\name{checkPeakMatrixOrientation}
\alias{checkPeakMatrixOrientation}
\title{Check is input/output data are in format features in rows, samples in columns. Check if class labels are provided for all samples.}
\usage{
checkPeakMatrixOrientation(df, classes)
}
\arguments{
\item{df}{peak matrix}

\item{classes}{Vector of class labels}
}
\value{
matrix in where samples are represented in columns and features in rows
}
\description{
Check is input/output data are in format features in rows, samples in columns. Check if class labels are provided for all samples.
}
