% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRSDplot.R
\name{doRSDplot}
\alias{doRSDplot}
\title{Plot violin plots from the doRSD fucntion output}
\usage{
doRSDplot(RSD, plotTitle = NULL, base_size = 12, subtitle = NULL)
}
\arguments{
\item{RSD}{OUtput of doRSD function}

\item{plotTitle}{Main tilte for the plot}

\item{base_size}{Font size for plot fonts}

\item{subtitle}{Subtitle to include in PCA plot}
}
\value{
Ggplot object with plot(s)
}
\description{
Plot violin plots from the doRSD fucntion output
}
