% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRSDtable.R
\name{doRSDtable}
\alias{doRSDtable}
\title{Plot violin plots from the doRSD fucntion output}
\usage{
doRSDtable(RSD, QC_label = "QC", Blank_label = "Blank")
}
\arguments{
\item{RSD}{OUtput of doRSD function}

\item{QC_label}{Label used for QC samples. If set to NULL, assumes that no QC samples are present in data set}

\item{Blank_label}{Label used for Blank samples}
}
\value{
Table of RSD% values per group
}
\description{
Plot violin plots from the doRSD fucntion output
}
