% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_peaks_by_fraction}
\alias{filter_peaks_by_fraction}
\title{Filter features by fraction of missing values}
\usage{
filter_peaks_by_fraction(
  df,
  min_frac,
  classes = NULL,
  method = "QC",
  qc_label = "QC",
  remove_peaks = TRUE
)
}
\arguments{
\item{df}{peak intensity matrix}

\item{min_frac}{threshold of fraction of detection}

\item{classes}{character vector of class labels. If input is 
`SummarizedExperiment` object, a column of `colData` slot.}

\item{method}{method to use. 'QC' - withing QC samples, within' - within 
each sample class or across' - across all samples}

\item{qc_label}{class label for QC sample}

\item{remove_peaks}{remove filtered features from peak matrix or not}
}
\value{
'SummarizedExperiment' class object or list of filtered peak 
intensity matrix and matrix with flags.
}
\description{
Metabolomics datasets often contain 'features' with irreproducible peak 
intensity values, or with large numbers of missing values. This tool 
facilitates the remove of such features from a data matrix, based upon the 
relative proportion (minimum fraction) of samples containing non-missing 
values.
}
\examples{
df <- MTBLS79[ , MTBLS79$Batch==1]
df$Class[1:2] <- "Blank"
out <- filter_peaks_by_fraction(df=df, min_frac=1, 
    classes=df$Class, method='QC', qc_label='QC')
    
out <- filter_peaks_by_fraction(df=df, min_frac=1, 
    classes=df$Class, method='across', qc_label='QC')

out <- filter_peaks_by_fraction(df=df, min_frac=1, 
    classes=df$Class, method='within', qc_label='QC')

}
