% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes_methods.R
\docType{methods}
\name{getGlogLambdaOptimisationPlot,GlogOutput-method}
\alias{getGlogLambdaOptimisationPlot,GlogOutput-method}
\alias{getGlogLambdaOptimisationPlot}
\title{Method to get lamba optimisation plot from 'GlogOutput' class obejct}
\usage{
\S4method{getGlogLambdaOptimisationPlot}{GlogOutput}(glog_object)
}
\arguments{
\item{glog_object}{R object of 'GlogOuput' class}
}
\value{
'ggplot' class object
}
\description{
Method to get lamba optimisation plot from 'GlogOutput' class obejct
}
