% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes_methods.R
\docType{methods}
\name{getScaledPeakMatrix,GlogOutput-method}
\alias{getScaledPeakMatrix,GlogOutput-method}
\alias{getScaledPeakMatrix}
\title{Method to get peak matrix from 'GlogOutput' class object}
\usage{
\S4method{getScaledPeakMatrix}{GlogOutput}(glog_object)
}
\arguments{
\item{glog_object}{R object of 'GlogOuput' class}
}
\value{
peak matrix after glog scaling
}
\description{
Method to get peak matrix from 'GlogOutput' class object
}
