% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_transformation.R
\name{glog}
\alias{glog}
\title{variance stabilising (extended) generalised logarithm transformation}
\usage{
glog(y, y0 = 0, lambda)
}
\arguments{
\item{y}{values to be tranformed}

\item{y0}{offset applied to y (default = 0).}

\item{lambda}{transform parameter}
}
\value{
vector of transformed values
}
\description{
https://doi.org/10.1186/1471-2105-8-234
}
