% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_transformation.R
\name{glog_omptimise_lambda}
\alias{glog_omptimise_lambda}
\title{Search for optimal value of lambda by minimasing SSE.}
\usage{
glog_omptimise_lambda(upper_lim, df_qc)
}
\arguments{
\item{upper_lim}{upper limit to use for optimisation}

\item{df_qc}{peak matrix of QC samples}
}
\value{
optimised glog lambda value
}
\description{
Search for optimal value of lambda by minimasing SSE.
}
