% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_transformation.R
\name{glog_plot_optimised_labmda}
\alias{glog_plot_optimised_labmda}
\title{Plot SSE error of lamba optimisation process}
\usage{
glog_plot_optimised_labmda(optimised_lambda = NA, data_qc, upper_lim)
}
\arguments{
\item{optimised_lambda}{optimised lamba value from 'glog_optimise_output'}

\item{data_qc}{peak intenisty matrix of QC samples}

\item{upper_lim}{limit of the upper lambda value}
}
\value{
ggplot object containing optimisation plot
}
\description{
Plot SSE error of lamba optimisation process
}
