% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_transformation.R
\name{glog_plot_optimised_lambda}
\alias{glog_plot_optimised_lambda}
\title{Plot SSE error of lambda optimisation process}
\usage{
glog_plot_optimised_lambda(optimised_lambda = NA, data_qc)
}
\arguments{
\item{optimised_lambda}{optimised lambda value from 'glog_optimise_output'}

\item{data_qc}{peak intenisty matrix of QC samples}
}
\value{
ggplot object containing optimisation plot
}
\description{
Plot SSE error of lambda optimisation process
}
