% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_transformation.R
\name{glog_rescale_data}
\alias{glog_rescale_data}
\title{If glog optimisation fails, this function will scale values in the 
peak matrix to the 1/mean(total signal) over all samples.}
\usage{
glog_rescale_data(df)
}
\arguments{
\item{df}{peak intensity matrix}
}
\value{
scaled peak matrix
}
\description{
If glog optimisation fails, this function will scale values in the 
peak matrix to the 1/mean(total signal) over all samples.
}
