% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_transformation.R
\name{glog_transformation}
\alias{glog_transformation}
\title{Performs glog transformation on the data set. Using QC samples to estimate lamda and alpha parameters.}
\usage{
glog_transformation(df, classes, qc_label)
}
\arguments{
\item{df}{Data frame.}

\item{classes}{Vector of sample classes.}

\item{qc_label}{Lable for QC sample.}
}
\description{
Performs glog transformation on the data set. Using QC samples to estimate lamda and alpha parameters.
}
