% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes_methods.R
\docType{class}
\name{GlogOutput-class}
\alias{GlogOutput-class}
\alias{GlogOutput}
\title{S4 class for Variance stabilising generalised lograrithm transformation
 
This class defines structure and methods of glog_transformation output.}
\description{
S4 class for Variance stabilising generalised lograrithm transformation
 
This class defines structure and methods of glog_transformation output.
}
\section{Slots}{

\describe{
\item{\code{scaled_peak_matrix}}{a data frame, peak inntensity matrix after glog
transformation}

\item{\code{lambda_summary}}{a list containing optimised lambda value and if 
optimisation has failed}

\item{\code{lambda_optimisation_plot}}{ggplot2 object of glog lambda
optimisation}
}}

