% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_imputation.R
\name{impute_mode}
\alias{impute_mode}
\title{Imput missing values using mode of each feature}
\usage{
impute_mode(df, method)
}
\arguments{
\item{df}{a peak matrix with features in the rows, samples in the columns}

\item{method}{mode to use for missing value imputation. 'mn' for mean and 
'md'for median.}
}
\value{
data frame of missing value imputed peak intensity matrix
}
\description{
Imput missing values using mode of each feature
}
