% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Plot multiple ggplot objects}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{List of ggplot objects}

\item{file}{name of output pdf files}

\item{cols}{number of columns}

\item{layout}{layout matrix}
}
\description{
Plot multiple ggplot objects
}
