% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{normalise_to_sum}
\alias{normalise_to_sum}
\title{Normalise peak table to the total sum of peak intensities}
\usage{
normalise_to_sum(df)
}
\arguments{
\item{df}{Data frame.}
}
\value{
Normalised peak matrix.
}
\description{
Normalise peak table to the total sum of peak intensities
}
