% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{pqn_normalisation_orig}
\alias{pqn_normalisation_orig}
\title{PQN normalisation}
\usage{
pqn_normalisation_orig(df, classes, qc_label)
}
\arguments{
\item{df}{Data frame.}

\item{classes}{Vector of class labels.}

\item{qc_label}{Label used for QC samples.}
}
\value{
List of normalised data set and correction coefficients
}
\description{
PQN normalisation
}
