% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_main.R
\name{QCRSC}
\alias{QCRSC}
\title{Quality Control-Robust Spline Correction (QC-RSC)}
\usage{
QCRSC(
  df,
  order,
  batch,
  classes,
  spar = 0,
  log = TRUE,
  minQC = 5,
  qc_label = "QC"
)
}
\arguments{
\item{df}{A data frame of values to be corrected (samples in columns and
features in rows).}

\item{order}{A numeric vector indicating the order in which samples
were measured.}

\item{batch}{A vector indicating the batch each sample was measured in.
If only one batch then all values should be set to 1.}

\item{classes}{A factor or character vector of sample classes. All QC
samples should be labelled  'QC'.}

\item{spar}{Spline smoothing parameter. Should be in the range 0 to 1.
If set to 0 it will be estimated using leave-one-out cross-validation.}

\item{log}{TRUE or FALSE to perform the signal correction fit on the log
scaled data. Default is TRUE.}

\item{minQC}{Minimum number of QC samples required for signal correction.}

\item{qc_label}{Class label for QC sample.}
}
\value{
A data frame of corrected values.
}
\description{
Implementation of signal correction algorithm as described
in Kirwan et al, Anal. Bioanal. Chem., 405 (15), 2013
}
\details{
The smoothing parameter (spar) can be optimised using leave-one-out cross
validation to avoid overfitting.
}
\examples{
classes <- MTBLS79$Class
batch <- MTBLS79$Batch
order <- c(1:ncol(MTBLS79))

out <- QCRSC(df = MTBLS79[1:10, ], order = order, batch = MTBLS79$Batch,
classes = MTBLS79$Class, spar = 0, minQC = 4)

}
\references{
Kirwan et al, Anal. Bioanal. Chem., 405 (15), 2013
\url{https://dx.doi.org/10.1007/s00216-013-6856-7}
}
\author{
Andris Jankevics \email{a.jankevics@bham.ac.uk}
}
