% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{remove_peaks}
\alias{remove_peaks}
\title{Remove features from peak intensity matrix}
\usage{
remove_peaks(df, rem_index)
}
\arguments{
\item{df}{Peak intensity matrix}

\item{rem_index}{Logical vector containing TRUE vales for features to remove}
}
\value{
data frame, filtered peak intensity matrix
}
\description{
Filter to remove features
}
\examples{
attach (testData)
rem_index <- testData$remove_peaks$rem_index

out <- remove_peaks(df=testData$data, rem_index = rem_index)

}
