% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderFactor.R
\name{reorderFactorLevels}
\alias{reorderFactorLevels}
\title{Function to create sorted class labels from claaa names}
\usage{
reorderFactorLevels(labels, QC_label = "QC", Blank_label = "Blank")
}
\arguments{
\item{labels}{Vector of class labels.}

\item{QC_label}{Label used for QC samples. If set to NULL, assumes that no QC samples are present in data set}

\item{Blank_label}{Label used for blank samples, if set to NULL no samples will be removed}
}
\value{
Vector of reordered class labels
}
\description{
Function to create sorted class labels from claaa names
}
