% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_imputation.R
\name{replace_na}
\alias{replace_na}
\title{Replace missing values for specific peak matrix feature}
\usage{
replace_na(x, df, vals)
}
\arguments{
\item{x}{index of matrix feature}

\item{df}{a peak matrix with features in the rows, samples in the columns}

\item{vals}{vector object with the same length as number of features in 
peak matrix to replca NA values with}
}
\value{
feature at specified index 'x' with replaced missing values
}
\description{
Replace missing values for specific peak matrix feature
}
