% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbcmsPlot.R
\name{sbcmsPlot}
\alias{sbcmsPlot}
\title{Plot the output from signal batch correction for the selected or
the first 100 features}
\usage{
sbcmsPlot(
  df,
  corrected_df,
  classes,
  batch,
  indexes = NULL,
  qc_label = "QC",
  output = "sbcms_plots.pdf"
)
}
\arguments{
\item{df}{A data frame containing the original data before correction
(samples in columns, features in rows).}

\item{corrected_df}{A data frame containing the corrected data,
as output by QCRSC}

\item{classes}{A factor or character vector of sample classes.
All QC samples should be labelled 'QC'.}

\item{batch}{A vector indicating the batch each sample was measured in.
If only one batch was measured then all values should be set to 1}

\item{indexes}{Numeric vector defining whihc features from data frame to 
plot. If set to NULL will plot the first 100.}

\item{qc_label}{Class label for QC sample.}

\item{output}{Filename of the output pdf file. Can include the path.
If set to NULL output will be list object containing ggplot2 plots.}
}
\value{
Pdf file or list object showing data before and after 
signal correction
}
\description{
Plot the output from signal batch correction for the selected or
the first 100 features
}
\examples{

classes <- MTBLS79$Class
batch <- MTBLS79$Batch
order <- c(1:ncol(MTBLS79))
data <- SummarizedExperiment::assay(MTBLS79[1:10, ] )

out <- QCRSC(d =data, order=order, batch=batch, classes=classes,
    spar=0, minQC=4)
plots <- sbcmsPlot (df=data, corrected_df=out, classes, batch,
    output=NULL) 

}
