% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme_pub.R
\name{scale_colour_Publication}
\alias{scale_colour_Publication}
\title{Ggplot colors}
\usage{
scale_colour_Publication(...)
}
\arguments{
\item{...}{No parameters to specify}
}
\description{
Ggplot colors
}
