% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme_pub.R
\name{scale_fill_Publication}
\alias{scale_fill_Publication}
\title{Ggplot fill colors}
\usage{
scale_fill_Publication(...)
}
\arguments{
\item{...}{No parameters to specify}
}
\description{
This ggplot2 theme is modified from original by  Koundinya Desiraju.
Availaible from here: https://rpubs.com/Koundy/71792
}
