% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme_pub.R
\name{set_scale}
\alias{set_scale}
\title{Ggplot color scales function}
\usage{
set_scale(y = NULL, name = "PCA", ...)
}
\arguments{
\item{y}{vector or classes}

\item{name}{type of plotted object}

\item{...}{Other parameters}
}
\description{
Ggplot color scales function
}
