% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes_methods.R
\name{setGlogLambdaOptimisationPlot,GlogOutput-method}
\alias{setGlogLambdaOptimisationPlot,GlogOutput-method}
\alias{setGlogLambdaOptimisationPlot}
\title{Method to set lamba optimisation plot slot}
\usage{
\S4method{setGlogLambdaOptimisationPlot}{GlogOutput}(glog_object, lambda_optimisation_plot)
}
\arguments{
\item{glog_object}{R object of 'GlogOuput' class}

\item{lambda_optimisation_plot}{output from 'glog_plot_optimised_labmda'
transformation}
}
\value{
GlogObject class object with updated plot slot
}
\description{
Method to set lamba optimisation plot slot
}
