% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes_methods.R
\name{setGlogLambdaSummary,GlogOutput-method}
\alias{setGlogLambdaSummary,GlogOutput-method}
\alias{setGlogLambdaSummary}
\title{Method to set glog lambda optimisation summary}
\usage{
\S4method{setGlogLambdaSummary}{GlogOutput}(glog_object, lambda, lambda_opt, error_flag)
}
\arguments{
\item{glog_object}{R object of 'GlogOuput' class}

\item{lambda}{glog lambda value}

\item{lambda_opt}{glog optimised lambda value}

\item{error_flag}{logical value from SSE optimisation call
transformation}
}
\value{
GlogObject class object with updated lambda summary slot
}
\description{
Method to set glog lambda optimisation summary
}
