% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes_methods.R
\docType{methods}
\name{setScaledPeakMatrix,GlogOutput-method}
\alias{setScaledPeakMatrix,GlogOutput-method}
\alias{setScaledPeakMatrix}
\title{Method to set peak matrix slot after glog scaling}
\usage{
\S4method{setScaledPeakMatrix}{GlogOutput}(glog_object, scaled_peak_matrix)
}
\arguments{
\item{glog_object}{R object of 'GlogOuput' class}

\item{scaled_peak_matrix}{a data frame, peak inntensity matrix after glog
transformation}
}
\value{
GlogObject class object with updated peak matrix slot
}
\description{
Method to set peak matrix slot after glog scaling
}
