% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{smu_normalisation}
\alias{smu_normalisation}
\title{SMU normalisation}
\usage{
smu_normalisation(df)
}
\arguments{
\item{df}{Data frame.}
}
\value{
Normalised data frame.
}
\description{
SMU normalisation
}
