% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_transformation.R
\name{SSE}
\alias{SSE}
\title{Internal glog function}
\usage{
SSE(lambda, alpha, y)
}
\arguments{
\item{lambda}{lambda.}

\item{alpha}{alpha.}

\item{y}{values.}
}
\description{
Internal glog function
}
