% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testData}
\alias{testData}
\title{Testing data set}
\format{List of testing data sets}
\source{
https://www.ebi.ac.uk/metabolights/MTBLS79
}
\usage{
testData
}
\description{
Subset of DIMS data set to test package functions
}
\keyword{datasets}
