% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme_pub.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Ggplog theme, source ...}
\usage{
theme_Publication(base_size = 14)
}
\arguments{
\item{base_size}{Ggplot font size}
}
\description{
Ggplog theme, source ...
}
