% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPeakMatrix.R
\name{check_input_data}
\alias{check_input_data}
\title{Check if input data is object of 'SummarizedExperiment',
if not convert inputs into 'SummarizedExperiment' container.}
\usage{
check_input_data(df, classes = NULL)
}
\arguments{
\item{df}{A matrix-like object (e.g. and ordinary matrix, a data frame) of 
peak intensities.}

\item{classes}{vector of class labels.}
}
\value{
object of class of 'SummarizedExperiment'.
}
\description{
Check if input data is object of 'SummarizedExperiment',
if not convert inputs into 'SummarizedExperiment' container.
}
