% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPeakMatrix.R
\name{check_input_data}
\alias{check_input_data}
\title{If needed convert input data to object of class 
\link[SummarizedExperiment]{SummarizedExperiment}}
\usage{
check_input_data(df, classes = NULL)
}
\arguments{
\item{df}{A matrix-like (e.g. an ordinary matrix, a data frame) object with
all values of class \code{numeric()} or \code{integer()} of peak
intensities, areas or other quantitative characteristic.}

\item{classes}{\code{character()}, vector of class labels. Must be the same 
length as the number of sample in the input peak table.}
}
\value{
object of class of 
\link[SummarizedExperiment]{SummarizedExperiment}.
}
\description{
If needed convert input data to object of class 
\link[SummarizedExperiment]{SummarizedExperiment}
}
