% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPeakMatrix.R
\name{check_peak_matrix}
\alias{check_peak_matrix}
\title{Check if peak matrix is in format features in rows, samples in columns 
and that all cells contain numeric values.}
\usage{
check_peak_matrix(peak_data, classes = NULL)
}
\arguments{
\item{peak_data}{peak matrix}

\item{classes}{vector of class labels}
}
\value{
matrix where samples are represented in columns and features in rows
}
\description{
All functions in pmp pacakge expect input peak matrix to have samples 
as columns and measured features in rows. This function will check input 
matrix orientation and will transpose it if needed. If class labels are 
provided this function will check if the length of labels matches number 
of samples in peak matrix.
}
