% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPeakMatrixOrientation.R
\name{check_peak_matrix_orientation}
\alias{check_peak_matrix_orientation}
\title{Check is input/output data are in format features in rows, samples in columns. Check if class labels are provided for all samples.}
\usage{
check_peak_matrix_orientation(peak_data, classes = NULL)
}
\arguments{
\item{peak_data}{peak matrix}

\item{classes}{Vector of class labels}
}
\value{
matrix where samples are represented in columns and features in rows
}
\description{
Check is input/output data are in format features in rows, samples in columns. Check if class labels are provided for all samples.
}
