% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_peaks_by_blank}
\alias{filter_peaks_by_blank}
\title{Filter peaks by blank}
\usage{
filter_peaks_by_blank(
  df,
  fold_change,
  classes,
  blank_label,
  qc_label = NULL,
  remove_samples = TRUE,
  remove_peaks = TRUE,
  fraction_in_blank = 0
)
}
\arguments{
\item{df}{peak intensity matrix}

\item{fold_change}{minimum fold change between analytical and blank samples.}

\item{classes}{character vector of class labels. If input is 
`SummarizedExperiment` object, a column of `colData` slot.}

\item{blank_label}{class label used to identify blank samples}

\item{qc_label}{class label for QC sample. If not NULL will use QC samples 
to calculate the mean intensity}

\item{remove_samples}{remove blank samples from peak matrix or not}

\item{remove_peaks}{remove filtered features from peak matrix or not}

\item{fraction_in_blank}{number between 0 to 1 to specify fraction in how 
many blanks peaks should be present}
}
\value{
'SummarizedExperiment' class object or list of filtered peak 
intensity matrix and matrix with flags.
}
\description{
Metabolomics datasets often contain many 'features' of non-biological 
origin e.g. those associated with extraction and analysis solvents. 
This tool facilitates the removal of such features from the data matrix, 
as defined using an appropriate blank' sample.
}
\examples{
df <- MTBLS79[ , MTBLS79$Batch==1]
df$Class[1:2] <- "Blank"
out <- filter_peaks_by_blank(df=df, fold_change=1.2, 
   classes=df$Class, blank_label="Blank", qc_label=NULL, 
   remove_samples=FALSE, remove_peaks=TRUE, fraction_in_blank=0)

}
