% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_peaks_by_blank}
\alias{filter_peaks_by_blank}
\title{Blank filter}
\usage{
filter_peaks_by_blank(df, fold_change, classes, blank_label, qc_label = NULL,
  remove = FALSE)
}
\arguments{
\item{df}{Input data frame.}

\item{fold_change}{Threshold of fold change for blank filter.}

\item{classes}{Vector of class labels.}

\item{blank_label}{Class label for blank samples}

\item{qc_label}{Class label for QC sample}

\item{remove}{If features abobe threshold shoule be removed or not}
}
\description{
Blank filter
}
