% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_peaks_by_fraction}
\alias{filter_peaks_by_fraction}
\title{Filter features by fraction of missing values}
\usage{
filter_peaks_by_fraction(
  df,
  min_frac,
  classes = NULL,
  method = "QC",
  qc_label = "QC"
)
}
\arguments{
\item{df}{peak intensity matrix}

\item{min_frac}{threshold of fraction of detection}

\item{classes}{vector of class labels}

\item{method}{method to use. 'QC' - withing QC samples, within' - within 
each sample class or across' - across all samples}

\item{qc_label}{class label for QC sample}
}
\value{
list of filtered peak intensity matrix and matrix with flags
}
\description{
Metabolomics datasets often contain 'features' with irreproducible peak 
intensity values, or with large numbers of missing values. This tool 
facilitates the remove of such features from a data matrix, based upon the 
relative proportion (minimum fraction) of samples containing non-missing 
values.
}
\examples{
out <- filter_peaks_by_fraction(df=pmp:::testData$data, min_frac=1, 
    classes=pmp:::testData$class, method='QC', qc_label='QC')
    
out <- filter_peaks_by_fraction(df=pmp:::testData$data, min_frac=1, 
    classes=pmp:::testData$class, method='across', qc_label='QC')

out <- filter_peaks_by_fraction(df=pmp:::testData$data, min_frac=1, 
    classes=pmp:::testData$class, method='within', qc_label='QC')

}
