% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_peaks_by_fraction}
\alias{filter_peaks_by_fraction}
\title{Filter features by fraction of missing values}
\usage{
filter_peaks_by_fraction(df, min_frac, classes = NULL, method = "QC",
  qc_label = "QC")
}
\arguments{
\item{df}{Peak intensity matrix}

\item{min_frac}{Threshold of fraction of detection}

\item{classes}{Vector of class labels}

\item{method}{Method to use, default is QC}

\item{qc_label}{Class label for QC sample}
}
\description{
Filter features by fraction of missing values
}
