% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_peaks_by_rsd}
\alias{filter_peaks_by_rsd}
\title{Filter features by RSD% of QC samples}
\usage{
filter_peaks_by_rsd(df, max_rsd, classes, qc_label)
}
\arguments{
\item{df}{Peak intensity matrix}

\item{max_rsd}{Threshold of QC RSD% value}

\item{classes}{Vector of class labels}

\item{qc_label}{Class label for QC sample}
}
\description{
Filter features by RSD% of QC samples
}
