% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_transformation.R
\name{glog}
\alias{glog}
\title{Internal glog function}
\usage{
glog(y, alpha, lambda)
}
\arguments{
\item{y}{values.}

\item{alpha}{alpha.}

\item{lambda}{lambda.}
}
\description{
Internal glog function
}
