% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_transformation.R
\name{glog}
\alias{glog}
\title{variance stabilising generalised logarithm transformation}
\usage{
glog(y, alpha, lambda)
}
\arguments{
\item{y}{values to be tranformed}

\item{alpha}{alpha.}

\item{lambda}{transform parameter}
}
\description{
https://doi.org/10.1186/1471-2105-8-234
}
