% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_transformation.R
\name{glog_transformation}
\alias{glog_transformation}
\title{Performs glog transformation on the data set,
using QC samples to estimate lambda.}
\usage{
glog_transformation(df, classes, qc_label, lambda = NULL)
}
\arguments{
\item{df}{Peak intensity matrix}

\item{classes}{vector of class labels}

\item{qc_label}{class label for QC sample}

\item{lambda}{If not NULL will use provided value for glog lambda}
}
\value{
An \link[SummarizedExperiment]{SummarizedExperiment} object
or numeric object of original input data structure.
}
\description{
https://doi.org/10.1186/1471-2105-8-234
}
\examples{
attach (testData)
out <- mv_imputation(df=testData$data, method='knn')
out <- glog_transformation (df=out, classes=testData$class,
    qc_label='QC')

}
