% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_transformation.R
\name{glog_transformation}
\alias{glog_transformation}
\title{Performs glog transformation on the data set, 
using QC samples to estimate lambda.}
\usage{
glog_transformation(df, classes, qc_label)
}
\arguments{
\item{df}{Peak intensity matrix}

\item{classes}{Vector of class labels}

\item{qc_label}{Class label for QC sample}
}
\value{
data frame, peak inntensity matrix after glog transformation
}
\description{
https://doi.org/10.1186/1471-2105-8-234
}
\examples{
attach (testData)
out <- mv_imputation(df=testData$data, method = 'knn')
out <- glog_transformation (df=out, classes = testData$class,
 qc_label = 'QC')

}
