% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{normalise_to_sum}
\alias{normalise_to_sum}
\title{Normalise peak table to the total sum of peak intensities}
\usage{
normalise_to_sum(df, check_df = TRUE)
}
\arguments{
\item{df}{Data frame}

\item{check_df}{If set to TRUE will check if input data needs to be
transposed, so that features are in rows.}
}
\value{
Normalised peak matrix.
}
\description{
Normalise peak table to the total sum of peak intensities
}
\examples{
attach (testData)
out <- normalise_to_sum (testData$data)

}
