% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{pqn_normalisation}
\alias{pqn_normalisation}
\title{Normalise peak table using PQN method}
\usage{
pqn_normalisation(df, classes, qc_label)
}
\arguments{
\item{df}{data frame}

\item{classes}{vector of class labels}

\item{qc_label}{label used for QC samples. If set to 'all', all samples will
be used to calculate correction factor}
}
\value{
list of normalised data set and correction coefficients
}
\description{
Normalise peak table using PQN method
}
\examples{
attach (testData)
pqn_normalisation(df=testData$data, classes=testData$class, qc_label='QC')

}
