% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{pqn_normalisation}
\alias{pqn_normalisation}
\title{Normalise peak table using PQN method}
\usage{
pqn_normalisation(df, classes, qc_label, ref_mean = NULL)
}
\arguments{
\item{df}{data frame}

\item{classes}{vector of class labels}

\item{qc_label}{label used for QC samples. If set to 'all', all samples will
be used to calculate correction factor}

\item{ref_mean}{Vector of reference mean values to use instead of calculating
from QC sample group. If set to NULL, QC sample data will be used.}
}
\value{
list of normalised data set and correction coefficients
}
\description{
Normalise peak table using PQN method
}
\examples{

df <- MTBLS79[ , MTBLS79$Batch==1]
pqn_normalisation(df=df,
    classes=df$Class, qc_label='QC')

}
