% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{remove_peaks}
\alias{remove_peaks}
\title{Remove features from peak intensity matrix}
\usage{
remove_peaks(df, rem_index)
}
\arguments{
\item{df}{peak intensity matrix}

\item{rem_index}{logical vector containing TRUE vales for features to remove}
}
\value{
data frame, filtered peak intensity matrix
}
\description{
Filter to remove features.
}
\examples{
df <- MTBLS79[ , MTBLS79$Batch==1]
rem_index <- vector(mode="logical", 
    length=nrow(SummarizedExperiment::assay(df)))
rem_index[c(1, 20, 456, 789)] <- TRUE
out <- remove_peaks(df=df, rem_index=rem_index)

}
