% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPeakMatrix.R
\name{return_original_data_structure}
\alias{return_original_data_structure}
\title{If input data object was not of class of 
\link[SummarizedExperiment]{SummarizedExperiment},
convert output to the original R data structure.}
\usage{
return_original_data_structure(summarized_experiment_object)
}
\arguments{
\item{summarized_experiment_object}{object of class
\code{SummarizedExperiment}}
}
\value{
A matrix-like (e.g. an ordinary matrix, a data frame) object with
all values of class \code{numeric()} or \code{integer()}. 
Values from \code{colData} and \code{rowData} elements are not returned.
}
\description{
All values stored in \code{metadata} element of \code{SummarizedExperiment}
will be exported as \link[base]{attributes} of the output object.
}
