% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPeakMatrix.R
\name{return_original_data_structure}
\alias{return_original_data_structure}
\title{If input data were not of class of 'SummarizedExperiment',
convert output to the original R data structure type.}
\usage{
return_original_data_structure(peak_data)
}
\arguments{
\item{peak_data}{peak matrix}
}
\value{
peak matrix object of the same data structure as original input data
}
\description{
If input data were not of class of 'SummarizedExperiment',
convert output to the original R data structure type.
}
